Task3/Hands Sign/requirements.txt
streamlit
opencv-python-headless
mediapipe
numpy
tensorflow
Pillow
protobuf
matplotlib
scikit-learn